Fixup how LNT handles dates and times in the UI
LNT has always confused people by showing dates and times in UTC, but
without adding the UTC prefix on the end of the string.  In this commit,
make the server pass all dates as UTC ISO 8601 date time strings. In the
browser convert these to the user's time zone and locale. Also, in
recent results locations use relative times, with a tooltip for the full
time string.