Prevent building of libunwind, libcxx, libcxxabi, and lld on powerpc64 buildbots
libunwind, libcxx, libcxxabi, and lld are not supported (and do not work) on powerpc64 (neither BE nor LE). However, they are always included in the code that the sanitizer buildbots fetch and compile. Changes made to them sometimes cause the powerpc64 buildbots to stop working until someone (like me) goes in and corrects the compilation problems. Until they are supported they need to be removed from what the powerpc64 buildbots fetch and build.

Checking the BUILDBOT_BUILDERNAME environment variable in buildbot_functions.sh was the simplest way I could see to do this and also let me try out the changes locally on the powerpc64 buildbots before submitting this revision.

http://reviews.llvm.org/D10033